+++ spare/tests/test_cli.py
@@ -1,5 +1,9 @@
 from click.testing import CliRunner
-from spare.cli import create_cli, restore_cli, validate_cli
+from spare.cli import create_cli
+from spare.cli import lock_cli
+from spare.cli import restore_cli
+from spare.cli import unlock_cli
+from spare.cli import validate_cli
 
 
 def test_create_cli(endpoint, access_key, secret_key, temporary_path):
@@ -33,3 +37,43 @@ def test_create_cli(endpoint, access_key, secret_key, temporary_path):
         '--bucket', 'foobar',
     ])
     assert result.exit_code == 0
+
+    # lock once
+    result = runner.invoke(lock_cli, [
+        '--endpoint', endpoint.geturl(),
+        '--access-key', access_key,
+        '--secret-key', secret_key,
+        '--password', 'foobar',
+        '--bucket', 'foobar',
+    ])
+    assert result.exit_code == 0
+
+    # lock again
+    result = runner.invoke(lock_cli, [
+        '--endpoint', endpoint.geturl(),
+        '--access-key', access_key,
+        '--secret-key', secret_key,
+        '--password', 'foobar',
+        '--bucket', 'foobar',
+    ])
+    assert result.exit_code == 1
+
+    # unlock
+    result = runner.invoke(unlock_cli, [
+        '--endpoint', endpoint.geturl(),
+        '--access-key', access_key,
+        '--secret-key', secret_key,
+        '--password', 'foobar',
+        '--bucket', 'foobar',
+    ])
+    assert result.exit_code == 0
+
+    # unlock again
+    result = runner.invoke(unlock_cli, [
+        '--endpoint', endpoint.geturl(),
+        '--access-key', access_key,
+        '--secret-key', secret_key,
+        '--password', 'foobar',
+        '--bucket', 'foobar',
+    ])
+    assert result.exit_code == 1